refactor: 리뷰 반영  - LottoAnswer LottoAuto 클래스 생성 및 Lotto 클래스 상속  - LottoFactory 메서드별 리턴 클래스 Answer 와 Auto 로 분류
