--- src/main/java/lotto/domain/LottoFactory.java
@@ -1,11 +1,0 @@
-package lotto.domain;
-
-import java.util.List;
-
-public interface LottoFactory {
-
-    Lotto newInstance();
-
-    Lotto newInstance(List<Integer> numbers);
-
-}
