+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.lotto;
+
+import lotto.domain.LottoNumbers;
 
 import java.util.ArrayList;
 import java.util.Collections;
