+++ src/main/java/lotto/domain/lotto/LottoAuto.java
@@ -1,0 +1,14 @@
+package lotto.domain.lotto;
+
+import java.util.List;
+
+public class LottoAuto extends Lotto {
+
+    public LottoAuto() {
+    }
+
+    @Override
+    public List<Integer> getNumbers() {
+        return super.getNumbers();
+    }
+}
